backend/requirements.txt
# ==================== 核心框架 ====================
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# ==================== 数据库 ====================
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
redis>=5.0.0

# ==================== 认证 ====================
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.6
email-validator>=2.0.0

# ==================== 配置 ====================
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# ==================== LangGraph / LLM ====================
langgraph>=0.0.40
langchain-core>=0.1.0
langchain-openai>=0.0.5

# ==================== PPT 生成 ====================
python-pptx>=0.6.21

# ==================== HTTP 客户端 ====================
httpx>=0.26.0

# ==================== 文件处理 ====================
aiofiles>=23.2.1
